Implemented a new system and some small updates
Small update: Added a quit game feature

System: Added levels of difficulty to the game that can be set at the start of one. Player now manages money, a cap of stocks and a lifespan. Will add an upgrade to the system to be able to extend lifespan and increase cap